HELM-104: Create automated build for Docker Images
Create a workflow to build and publish a Docker image to DockerHub. The workflow provides the following steps on
CircleCi:

 1. Build and test from source
 2. Create a tarball of the build artefact
 3. Create a Docker Image with installing from the tarball artifact
 4. Only publish on Dockerhub when a git tag as version number is set

Steps 1-3 are triggered for every branch including pushing a git tag. Publish is only triggered when a git tag is
pushed. The version number from the git tag is used for the Docker Image and during the plugin build.

For any other branches, the tarball and a Docker Image is created but stored as artifact on CircleCi under the specific
build number and can be downloaded from there.

The dependency for Grafana version can be controlled in grafana.version. With this commit it is set to 5.2.4.

Added documentation how to build the Docker image manually from source.