normalize datasource calling conventions to match grafana's
The fault-ds took a non-standard `query` argument that was an
object rather than a string, which caused all kinds of weirdness
when trying to interact with it like other datasources.

The `metricFindQuery` method *does* take a second argument which
contains extra options so I refactored the fault-ds to always
take the attribute or query as the first arg, and then do the
attribute/operator/comparator logic based on a `queryType`
property on the options.  If `queryType` is missing, assume
it's a normal "values" query.

I also fixed the return values to have the right fields to show
properly in the `Preview of values` of the variable editor. They
were blank before.