HELM-133 entity datasource
This turns the fault datasource into a generic entity datasource.
To do so required a number of refactorings to make it easier to
"task-switch" between different entity types, currently alarm
and node.

First, I created NodeEntity and AlarmEntity classes to encapsulate
the logic to perform queries, transform results into a useful grafana
table result, etc.  This included pulling the `toTable` method out
of `datasource.js`.  It also includes things like "featured
attributes", column metadata, and so on.

I also changed a number of the mapping and UI classes to take
an entity so that they can be passed around properly to perform
various entity-specific behavior.

I updated the query control to choose between entities with a
select box, instead of just saying "SELECT all".