HELM-178: require Grafana 6
This also updates a number of places where code from Grafana's
upstream table was backported to more closely match the versions
in Grafana 6.3.x, and fixes a bug in the color style handling.
It is up-to-date as of Grafana 6.3.3.
(git commit 84d0a71b25a949fde79a7054241a28d2f2e25999)