HELM-164: angular hack to make sure the + button reinitializes
Because the plus button (`uiSegmentSrv.newPlusButton()`) is a
reference in the <metric-segment> directive, it doesn't get
reinitialized properly once we add a column and attempt to
reset things.  The fix is to remove the <metric-segment> from the
DOM, create a new button, and then recreate it and have Angular
re-compile it.