Documentation changes
Updates to the About and Installation sections. Added text to provide context for some of the information.
Note: In the About section, it would be good to remove "Figure 1" from the caption. I have found instructions on how to do this, but could use some help. Less urgent, I have been unable to find information on how to change the titles in the left-hand menu (e.g., change "Getting started" to "Getting started with Helm".) In Installation, I noticed that Yarn is no longer listed in the requirements. I didn't see anything in the history in the repo to indicate this had been removed.